+++ frontend/package.json
@@ -4,18 +4,19 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^4.0.7",
+    "@babel/runtime-corejs2": "7.28.4",
     "@popperjs/core": "^2.11.8",
     "@rjsf/bootstrap-4": "5.24.13",
     "@rjsf/core": "5.24.13",
     "@rjsf/utils": "5.24.13",
+    "@rjsf/validator-ajv8": "^5.24.13",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.5.2",
-    "@types/apollo-upload-client": "^17.0.0",
-    "@types/jest": "^29.5.14",
-    "@types/json2csv": "^5.0.3",
-    "@types/node": "^22.10.2",
+    "@types/apollo-upload-client": "^19.0.0",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.7.2",
     "@types/react": "^19.2.2",
     "@types/react-dom": "^19.2.2",
     "@types/react-jsonschema-form": "^1.7.4",
@@ -26,7 +27,6 @@
     "graphql": "^16.9.0",
     "humps": "^2.0.1",
     "json-schema": "^0.4.0",
-    "json2csv": "^6.0.0-alpha.2",
     "jwt-decode": "^4.0.0",
     "react": "^19.2.0",
     "react-bootstrap": "^2.10.10",
@@ -61,8 +61,8 @@
     ]
   },
   "resolutions": {
-    "@types/react": "^18.3.17",
-    "@types/react-dom": "^18.3.5"
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2"
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
