+++ frontend/src/components/crud/DisplayTableContainer.tsx
@@ -6,7 +6,23 @@ import { HeaderGroup, useTable, Column } from "react-table";
 import EntityAPIClient, {
   EntityResponse,
 } from "../../APIClients/EntityAPIClient";
-import { downloadCSV } from "../../utils/CSVUtils";
+// Simple CSV download utility
+const downloadCSV = (data: string, fileName: string): void => {
+  const byteOrderMark = "\uFEFF";
+  const csvContent = byteOrderMark + data;
+  const blob = new Blob([csvContent], {
+    type: "text/csv, charset=UTF-8",
+  });
+  const url = URL.createObjectURL(blob);
+
+  const a = document.createElement("a");
+  a.href = url;
+  a.download = fileName;
+  a.click();
+  setTimeout(() => {
+    URL.revokeObjectURL(url);
+  });
+};
 
 type EntityData = Omit<EntityResponse, "boolField"> & { boolField: string };
