+++ frontend/src/components/crud/SimpleEntityCreateForm.tsx
@@ -1,6 +1,9 @@
 import React, { useState } from "react";
 import { JSONSchema7 } from "json-schema";
 import { Form } from "@rjsf/bootstrap-4";
+import { IChangeEvent } from "@rjsf/core";
+import { RJSFSchema } from "@rjsf/utils";
+import AJV8Validator from "@rjsf/validator-ajv8";
 import SimpleEntityAPIClient, {
   SimpleEntityRequest,
   SimpleEntityResponse,
@@ -66,18 +69,23 @@ const SimpleEntityCreateForm = (): React.ReactElement => {
     return <p>Created! ✔️</p>;
   }
 
-  const onSubmit = async ({ formData }: { formData: SimpleEntityRequest }) => {
-    const result = await SimpleEntityAPIClient.create({ formData });
-    setData(result);
+  const onSubmit = async (data: IChangeEvent<SimpleEntityRequest>) => {
+    if (data.formData) {
+      const result = await SimpleEntityAPIClient.create({ formData: data.formData });
+      setData(result);
+    }
   };
   return (
     <Form
       formData={formFields}
-      schema={schema}
+      schema={schema as RJSFSchema}
       uiSchema={uiSchema}
-      onChange={({ formData }: { formData: SimpleEntityRequest }) =>
-        setFormFields(formData)
-      }
+      validator={AJV8Validator}
+      onChange={(data: IChangeEvent<SimpleEntityRequest>) => {
+        if (data.formData) {
+          setFormFields(data.formData);
+        }
+      }}
       onSubmit={onSubmit}
     />
   );
