+++ frontend/src/components/pages/HooksDemo/HooksDemo.tsx
@@ -56,8 +56,8 @@ const HooksDemo = (): React.ReactElement => {
   // useRef allows us to store mutable, persistent values between
   // renders. Here we store an array of HTMLDivElement refs so
   // we can scroll each section into view.
-  const sectionRefs = useRef<React.RefObject<HTMLDivElement>[]>(
-    new Array(EMOJIS.length).fill(null).map(() => React.createRef()),
+  const sectionRefs = useRef<React.RefObject<HTMLDivElement | null>[]>(
+    new Array(EMOJIS.length).fill(null).map(() => React.createRef<HTMLDivElement | null>()),
   );
 
   useEffect(
