+++ frontend/src/reportWebVitals.ts
@@ -1,11 +1,11 @@
 const reportWebVitals = (onPerfEntry?: (metric: unknown) => void): void => {
   if (onPerfEntry && onPerfEntry instanceof Function) {
-    import("web-vitals").then(({ onCLS, onFID, onFCP, onLCP, onTTFB }) => {
+    import("web-vitals").then(({ onCLS, onFCP, onLCP, onTTFB, onINP }) => {
       onCLS(onPerfEntry);
-      onFID(onPerfEntry);
       onFCP(onPerfEntry);
       onLCP(onPerfEntry);
       onTTFB(onPerfEntry);
+      onINP(onPerfEntry);
     });
   }
 };
